added: override for No Settings option
This is for users who need to reactivate the Settings shortcut on the Main Menu but can't invoke the Button Menu. It can even be used if Kiosk mode is enabled, offering an always-available override that can't be accidentally triggered.

1) Go into the Showcase view of any window.
2) Push down twice so that the thumbnails clear away leaving just the backdrop.
3) Input the following sequence: right, down, left, up.
4) The thumbnails should return and you'll now have a Settings shortcut again.

The trick behind this is that if at any point during the sequence you press the wrong direction, the sequence will break and the thumbnails will return. This ensures that random inputs from your kids shouldn't accidentally trigger the override.