added: Secondary hide animation for StripShow
Stripshow now has two consecutive hide motions. The first triggers after 3 seconds to leave the info bar. The second triggers after a further 3 seconds to clear the screen entirely. This works around the problem of not being able to clear the screen manually when there are two rows of icons.