added: Wrap option for Muliplex (on by default)
The default browsing behaviour for Multiplex is now to wrap the contents around a central focused item. This can be reverted to the previous icon panel style using the Wrap: On/Off toggle.

NOTE: There is a glitch at present that causes Wall view to open whenever this option is toggled. Simply re-select Multiplex and you'll be fine, and the problem won't reoccur unless you toggle again. Do not report this.